fixed something in particle filter

+++ 2020-2021/botCode/utils/conversions.cpp
@@ -1,3 +1,3 @@
 const double SECS_TO_MILLIS = 1000;
-double secs_to_millis(int secs) { return secs * SECS_TO_MILLIS; }
-double millis_to_secs(int millis) { return millis / SECS_TO_MILLIS; }
+double secs_to_millis(double secs) { return secs * SECS_TO_MILLIS; }
+double millis_to_secs(double millis) { return millis / SECS_TO_MILLIS; }
