+++ 2020-2021/botCode/utils/conversions.hpp
@@ -1,7 +1,7 @@
 #ifndef CONVERSIONS_H
 #define CONVERSIONS_H
 
-double millis_to_secs(int millis);
-double secs_to_millis(int secs);
+double millis_to_secs(double millis);
+double secs_to_millis(double secs);
 
 #endif
